floats replaced with decimals control for course setting traansaction create not done

+++ ExchangeApp.BL/Models/Currency/CurrencyCoursesListModel.cs
@@ -6,8 +6,8 @@ public record CurrencyCoursesListModel : ModelBase
 {
     public required string Code { get; set; }
     public required string PhotoUrl { get; set; }
-    public required float AverageCourseRate { get; set; }
-    public float? BuyRate { get; set; }
-    public float? SellRate { get; set; }
+    public required decimal AverageCourseRate { get; set; }
+    public decimal? BuyRate { get; set; }
+    public decimal? SellRate { get; set; }
     public required CurrencyState Status { get; set; }
 }
